Renaming `elementConfigurationSection` to `elementConfiguration`.
