+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfiguration/ElementConfigurationSection.test.tsx
