+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfiguration/GroupBy/FieldComponent.tsx
@@ -18,7 +18,7 @@ import * as React from 'react';
 import { Field, useFormikContext } from 'formik';
 import { useContext } from 'react';
 
-import FieldSelect from 'views/components/aggregationwizard/elementConfigurationSections/FieldSelect';
+import FieldSelect from 'views/components/aggregationwizard/elementConfiguration/FieldSelect';
 import FieldTypesContext from 'views/components/contexts/FieldTypesContext';
 import { WidgetConfigFormValues } from 'views/components/aggregationwizard/WidgetConfigForm';
