+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfiguration/configurationFields/BooleanField.tsx
